Create for_challenge1.py
Ring in the New Year! A New Year's Eve party doesn't feel complete without a countdown from 10 to 1.

Use a for loop and the range() function for a countdown.

The output should look like:

10
9
8
7
6
5
4
3
2
1
Happy New Year! 🥳

Counting down from 10 to 1 on each line and then a Happy New Year message at the end.